requirements.txt
streamlit
numpy
Pillow
opencv-python-headless # Essential for server deployment (like Streamlit Cloud)